Simplify collapsing javascript.
Always explicitly set or clear the "checked" property, regardless of its
initial value. This makes it possible for the following commit to default posts
to uncollapsed.

Additionally, use a fixed length above which to collapse, instead of a factor.
This is mostly because we can't rely on clientHeight being limited to the
max-height anymore.