Don't change gallery-item background-color on hover.
My intention with the hover effect was never to change the background-color,
just to have the same color around the image as without hover. Before and after
the transition, that was the case, but during the transition there were some
semi-transparent background colors. This commit fixes that so the hover effect
is less distracting.